// src/app/models/usuario.ts
class Usuario {
    id: any;
    uid: number;
    agencia:string;
    email: string;
    // lastconection:number;
    name: string;
    nomAgencia: string;
    nombre:string;
    // sinceto:number;
    status: number;
    usuario: string;
    // empresa: string;
    // empresaId: number;
}
export { Usuario };