// src/app/models/agencia.ts
export class Agencia {
    id: any;
    nomAgencia: string;
    agencia_logobanner: string;
    agencia_logodocs: string;
    agencia_pie: string;
    aoPass: string;
    aoUser: string;
    celular: number;
    ciudad: string;
    currency: string;
    dirAgencia: string;
    direccionFiscal: string;
    estado: string;
    fecha_registro: string;
    idAgencia: number;
    id_agencia: string;
    pais: string;
    razonSocial: string;
    redSocial: string;
    rfc: string;
    telefono: number;
    utilidad: number;
    servicios: string;
}